mcp_architecture/requirements.txt
huggingface_hub #for download of google hear model
fastmcp
dotenv
ollama
librosa
numpy
scikit-learn
joblib
streamlit>=1.31.0  # For MCP visualization interface